+++ app/api/ada/index.js
@@ -46,6 +46,8 @@ import {
   NotEnoughMoneyToSendError,
   NotAllowedToSendMoneyToSameAddressError,
   NotAllowedToSendMoneyToRedeemAddressError,
+  NotEnoughFundsForTransactionFeesError,
+  AllFundsAlreadyAtReceiverAddressError,
   IncorrectWalletPasswordError,
 } from './errors';
 import { LOVELACES_PER_ADA } from '../../config/numbersConfig';
@@ -215,6 +217,13 @@ export default class AdaApi {
       return _createTransactionFeeFromServerData(response);
     } catch (error) {
       Logger.error('CardanoClientApi::calculateTransactionFee error: ' + stringifyError(error));
+      // eslint-disable-next-line max-len
+      if (error.message.includes('not enough money on addresses which are not included in output addresses set')) {
+        throw new AllFundsAlreadyAtReceiverAddressError();
+      }
+      if (error.message.includes('not enough money')) {
+        throw new NotEnoughFundsForTransactionFeesError();
+      }
       throw new GenericApiError();
     }
   }
@@ -438,6 +447,17 @@ export default class AdaApi {
     // return null;
   }
 
+  async postponeUpdate() {
+    Logger.debug('CardanoClientApi::postponeUpdate called');
+    try {
+      const response = await ClientApi.postponeUpdate(tlsConfig);
+      Logger.debug('CardanoClientApi::postponeUpdate success: ' + stringifyData(response));
+    } catch (error) {
+      Logger.error('CardanoClientApi::postponeUpdate error: ' + stringifyError(error));
+      throw new GenericApiError();
+    }
+  }
+
   async applyUpdate() {
     Logger.debug('CardanoClientApi::applyUpdate called');
     try {
@@ -525,6 +545,7 @@ export default class AdaApi {
       throw new GenericApiError();
     }
   }
+
 }
 
 // ========== TRANSFORM SERVER DATA INTO FRONTEND MODELS =========
@@ -565,9 +586,10 @@ const _createTransactionFromServerData = action(
       description: ctmDescription || '',
       numberOfConfirmations: data.ctConfirmations,
       addresses: {
-        from: data.ctInputAddrs,
-        to: data.ctOutputAddrs,
+        from: data.ctInputs.map(address => address[0]),
+        to: data.ctOutputs.map(address => address[0]),
       },
+      condition: data.ctCondition,
     });
   }
 );