+++ app/stores/AppStore.js
@@ -77,7 +77,7 @@ export default class AppStore extends Store {
   @computed get currentTheme(): string {
     const { result } = this.getThemeRequest.execute();
     if (this.isCurrentThemeSet) return result;
-    return environment.isMainnet() ? THEMES.CARDANO : THEMES.LIGHT_BLUE; // default
+    return environment.isMainnet() ? THEMES.DARK_BLUE : THEMES.LIGHT_BLUE; // default
   }
 
   @computed get isCurrentThemeSet(): boolean {