+++ app/themes/daedalus/cardano.js
@@ -15,6 +15,8 @@ export default {
 
   '--theme-input-hint-font': 'SFUIDisplay-Regular',
 
+  '--theme-scrollbar-thumb-background': '#c8ccce',
+
   '--theme-ada-redemption-headline-color': '#5e6066',
   '--theme-ada-redemption-instructions-color': '#5e6066',
   '--theme-ada-redemption-success-overlay-background-color': 'rgba(44, 187, 105, 0.88)',
@@ -36,7 +38,7 @@ export default {
   '--theme-ada-redemption-disclaimer-button-border-color': '#fafbfc',
 
   '--theme-bordered-box-background-color': '#fafbfc',
-  '--theme-bordered-box-border-color': '#c6cdd6',
+  '--theme-bordered-box-border': '1px solid #c6cdd6',
   '--theme-bordered-box-text-color': '#5e6066',
 
   '--theme-button-primary-background-color': '#2cbb69',
@@ -45,38 +47,38 @@ export default {
   '--theme-button-primary-background-color-disabled': 'rgba(44, 187, 105, 0.3)',
   '--theme-button-primary-text-color-disabled': '#fafbfc',
   '--theme-button-primary-text-color': '#fafbfc',
+
   '--theme-button-flat-background-color': '#f0f3f5',
   '--theme-button-flat-background-color-hover': '#f3f5f7',
   '--theme-button-flat-background-color-active': '#ebeef0',
+  '--theme-button-flat-background-color-disabled': 'rgba(240, 243, 245, 0.3)',
+  '--theme-button-flat-text-color-disabled': '#5e6066',
   '--theme-button-flat-text-color': '#5e6066',
 
   '--theme-checkbox-label-color': '#5e6066',
   '--theme-checkbox-border-color': '#2cbb69',
   '--theme-checkbox-background-color-checked': '#2cbb69',
 
+  '--theme-select-arrow-color': '#c6cdd6',
+  '--theme-select-arrow-color-open': '#5e6066',
+  '--theme-select-background-color': '#fafbfc',
+  '--theme-select-option-highlight-color': '#edeeef',
+  '--theme-select-option-text-color': '#5e6066',
+  '--theme-select-options-border-color': '#c6cdd6',
+  '--theme-select-options-shadow': 'none',
+
   '--theme-choice-tabs-text-color': '#5e6066',
   '--theme-choice-tabs-text-color-active': '#5e6066',
   '--theme-choice-tabs-bottom-border-color-active': '#5e6066',
 
-  '--theme-dialog-input-label-color': '#5e6066',
-  '--theme-dialog-input-text-color': '#5e6066',
-  '--theme-dialog-input-right-floating-indicator-color': 'rgba(94, 96, 102, 0.5)',
-  '--theme-dialog-input-placeholder-color': 'rgba(94, 96, 102, 0.5)',
-  '--theme-dialog-input-error-color': '#ea4c5b',
-  '--theme-dialog-switch-label-color': '#5e6066',
   '--theme-dialog-choice-tabs-text-color': '#5e6066',
   '--theme-dialog-choice-tabs-text-color-active': '#5e6066',
   '--theme-dialog-choice-tabs-bottom-border-color-active': '#5e6066',
-  '--theme-dialog-checkbox-label-color': '#5e6066',
-  '--theme-dialog-big-button-background-color': '#f1f3f5',
-  '--theme-dialog-big-button-border-color': '#d7dce0',
+  '--theme-dialog-big-button-background-color': '#f5f5f5',
+  '--theme-dialog-big-button-border-color': '#e1e1e1',
   '--theme-dialog-big-button-label-color': '#5e6066',
   '--theme-dialog-big-button-description-color': '#5e6066',
   '--theme-dialog-title-color': '#5e6066',
-  '--theme-dialog-separation-border-color': '#dfe4e8',
-  '--theme-dialog-input-background-color': '#fafbfc',
-  '--theme-dialog-input-border-color': '#c6cdd6',
-  '--theme-dialog-input-focus-border-color': '#5e6066',
 
   '--theme-input-border-color': '#c6cdd6',
   '--theme-input-label-color': '#5e6066',
@@ -123,11 +125,13 @@ export default {
   '--theme-receive-qr-code-background-color': 'transparent',
   '--theme-receive-qr-code-foreground-color': '#000',
 
+  '--theme-send-confirmation-dialog-send-values-color': '#ea4c5b',
+
   '--theme-settings-body-background-color': '#efefef',
   '--theme-settings-pane-background-color': '#fafbfc',
-  '--theme-settings-pane-border-color': '#c6cdd6',
+  '--theme-settings-pane-border': '1px solid #c6cdd6',
   '--theme-settings-menu-box-background-color': '#fafbfc',
-  '--theme-settings-menu-box-border-color': '#c6cdd6',
+  '--theme-settings-menu-box-border': '1px solid #c6cdd6',
   '--theme-settings-menu-item-text-color': '#5e6066',
   '--theme-settings-menu-item-text-color-active': '#5e6066',
   '--theme-settings-menu-item-text-color-enabled-hover': '#5e6066',
@@ -155,8 +159,9 @@ export default {
   '--theme-staking-font-color-accent': '#5e6066',
   '--theme-staking-font-color-regular': '#5e6066',
 
+  '--theme-switch-background-color': '#2cbb69',
   '--theme-switch-background-color-checked': '#2cbb69',
-  '--theme-switch-thumb-color-checked': '#2cbb69',
+  '--theme-switch-thumb-color-checked': '#fff',
   '--theme-switch-label-color': '#5e6066',
 
   '--theme-test-environment-label-background-color': '#ab1700',
@@ -175,6 +180,10 @@ export default {
   '--theme-transactions-card-income-background-color': '#e6ebf2',
   '--theme-transactions-card-expend-background-color': '#f2e6e6',
   '--theme-transactions-card-exchange-background-color': '#f2e6e6',
+  '--theme-transactions-state-failed-background-color': '#bdc5ce',
+  '--theme-transactions-state-failed-text-color': '#bdc5ce',
+  '--theme-transactions-state-pending-background-color': '#bdc5ce',
+  '--theme-transactions-state-pending-stripes-color': '#b2bac2',
   '--theme-transactions-priority-color': '#fafbfc',
   '--theme-transactions-priority-low-background-color': '#ab1700',
   '--theme-transactions-priority-medium-background-color': '#e6aa00',
@@ -183,6 +192,7 @@ export default {
   '--theme-transactions-icon-type-expend-background-color': '#84a2d2',
   '--theme-transactions-icon-type-income-background-color': '#2dc06c',
   '--theme-transactions-icon-type-exchange-background-color': '#10aca4',
+  '--theme-transactions-icon-type-failed-background-color': '#ea4c5b',
 
   '--theme-uploader-text-color': '#5e6066',
   '--theme-uploader-border-color': '#c6cdd6',
@@ -198,7 +208,7 @@ export default {
   '--theme-icon-ada-redemption-certificate-color': '#9eabbb',
   '--theme-icon-ada-summary-wallet-amount-symbol-color': '#5e6066',
   '--theme-icon-ada-summary-wallet-pending-confirmation-symbol-color': '#5e6066',
-  '--theme-icon-add-wallet-dialog-big-button-color': '#8a98ab',
+  '--theme-icon-add-wallet-dialog-big-button-color': '#ababab',
   '--theme-icon-copy-address-color': '#5e6066',
   '--theme-icon-back-button-color': '#5e6066',
   '--theme-icon-close-button-color': '#5e6066',
@@ -234,8 +244,8 @@ export default {
 
   '--theme-loading-background-color': '#ebeff2',
 
-  '--theme-mnemonic-background-color': '#e6e9ee',
-  '--theme-mnemonic-background-color-hover': '#eaecf1',
+  '--theme-mnemonic-background-color': '#f0f3f5',
+  '--theme-mnemonic-background-color-hover': '#f3f5f7',
 
   '--theme-send-logs-text-color': '#5e6066',